Adding contacts xls (snapshot)
This xls file is of course public-domain information -- it's not
something we compiled, in other words. In order to maintain authorship
information to give credit where it's due, we're leaving this as an xls,
rather than using a different format. We also added a link to the
canonical version, which can be found at
www.foia.gov/full-foia-contacts.xls as of this writing.